New Dists, updated Chi Squared
Added Bernouilli, Rademacher Distributions and added a rudimentary
probability method for Chi Squared. I won't be touching Chi Squared
stuff again for a while, there has to be a cleaner way to implement than
a table.

+++ StatsLib/Interfaces/IBinomial.cs
@@ -10,6 +10,6 @@
         /// <summary>
         /// The amount of individuals in the sample
         /// </summary>
-        double PopulationSize { get; }
+        uint PopulationSize { get; }
     }
 }
